Implemented export. Issue #15.
This is accessed through the HTML interface. Search results have a footer with some links like:

[HTML | Word]

When the user clicks on the "Word" link, he gets the browser's "Save As.." dialog and so can choose to Save (or just Open) the word doc wherever he desires.

The export functionality has been removed from the Flex App.

import & export collections no longer exist in the eXist DB.