Refactored the way we deploy files to eXist when installing.
Now we have an ANT script: exist/build.xml which is executed using exist/build.bat.

This builds the transcript.jar (eXist extension module) and places it in the correct location in the eXist installation.

Also copies Saxon jars and updates configuration files in eXist installation (so that the XSLT 2.0 code runs ok).

So the whole of the configuration of eXist is performed by this script. However, the database files (assets, reference.xml, xqueries, xslts) still need to be added manually.