Added xquery to count the number of instances of a concept (additional concepts) in all transcripts.
So now we bring up a dialog box if there is at least one instance about to be renamed - so the user can abort if necessary.

Also, "remove concept" is no longer allowed if at least one instance exists.

Also fixed situation where renaming to a name that is the same as the original name.