Major change to the way we use collections, resources and ids.
The collection hierarchy, resource filenames, and format of session/event ids is now transparent to the Flex app. Resource paths are now only relevant when storing resources (at which time the XQuery code works out some appropriate collection and filename). After storing resources, all queries use standard XPath expressions to locate data (completely ignoring the collection/resource path structure).

Session IDs no longer have to have the eventId as a prefix (although this is the default when the session id is created). eventId is now stored as an attribute in the session element.

Rewrote the event-selector part of the ImportWizard. Completely removed old "Open..." dialog so that it uses the same event selector as the Import wizard and then (if appropriate) brings up a session selector after that.

This code has not been fully tested so there will probably be a few more updates for this.

Related issues: #58, #83, #85, #86, #87