Fixed bug introduced in r430.
r430 resulted in the WordML being parsed IGNORING WHITESPACE. This results in some words being concatenated together.

To workaround this, we call the XQuery just after setting XML.ignoreWhitespace = false. Then we revert it to its original value as soon as the xquery returns (succeeds or fails).

This seems a little hacky. Maybe there is a way to do this without setting ignoreWhitespace to false?  Is there some attibute in WordML that gives us the information we need?